Use typeinformation from s:placed_signs in RedefineSign
The structure of the s:placed_signs dictionary appears to be
	`s:placed_signs[type][buffer][line]`
where type is either 'file' or 'project', buffer the buffer id and line
the linenumber of the placed sign. However, neomake#signs#RedefineSign
did only iterate over the first two levels of the hash, resulting in an
attempt to use a dictionary as an string/sign_id, which fails with:

E15: Invalid expression: 'sign place '.sign_id.' name=neomake_invisible buffer='.buf

This commit should fix that issue.